requirements.txt
fastapi==0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.1
google-cloud-secret-manager>=2.22.0
google-cloud-firestore>=2.16.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-json-logger==2.0.7
google-adk==1.2.1
